+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Fixing pytak installer and freezing bugs.
 
 - Adding pytak to the Misc. Dependencies for all operating systems
 - Disabling auto connect to TAK server in the FISSURE config file to prevent freezing without a TAK server
+- Unmerging IQEngine and TAK server installer items for Ubuntu 24.04
 
 ## 2025-12-12